separated RealTimeAtmosphere and StandardAtmosphere from Atmosphere abstract class. Validated StandardAtmosphere outputs with external data in unit test
